requirements.txt
flet
reportlab
pandas
pywin32
python-dotenv
psycopg2
openpyxl
